# web/openapi.yaml
openapi: 3.0.3
info:
  title: Xray Checker API
  description: API for monitoring Xray proxy status and configuration
  version: "1.0"
servers:
  - url: /
security:
  - basicAuth: []
paths:
...
  /api/v1/proxies/{stableID}:
    get:
      summary: Get proxy by ID
      description: Returns information for a specific proxy
      tags:
        - Proxies
      parameters:
        - name: stableID
          in: path
          required: true
          schema:
            type: string
          description: Proxy Stable ID (16-character hash)
      responses:
        '200':
          description: Proxy details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProxyInfo'
        '404':
          description: Proxy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorResponse'
...
  /api/v1/system/ip:
    get:
      summary: Get current IP
      description: Returns the current detected IP address
      tags:
        - System
      responses:
        '200':
          description: Current IP address
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SystemIPResponse'
        '500':
          description: Failed to get IP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorResponse'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response data (varies by endpoint)
        error:
          type: string
          description: Error message (only present on failure)
    APIErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Proxy not found"
...
    ProxyInfo:
      type: object
      properties:
        index:
          type: integer
          example: 0
        stableId:
          type: string
          example: "a1b2c3d4e5f67890"
        name:
          type: string
          example: "US-Server-1"
        subName:
          type: string
          description: Subscription name this proxy belongs to
          example: "Premium VPN"
        server:
          type: string
          example: "192.168.1.1"
        port:
          type: integer
          example: 443
        protocol:
          type: string
          enum: [vless, vmess, trojan, shadowsocks]
          example: "vless"
        proxyPort:
          type: integer
          example: 10000
        online:
          type: boolean
          example: true
        latencyMs:
          type: integer
          format: int64
          example: 150
...
    SystemIPResponse:
      type: object
      properties:
        ip:
          type: string
          example: "203.0.113.1"